Bugfixes and updating documentation
