+++ baseballcv/functions/utils/savant_utils/__init__.py
@@ -1,4 +1,4 @@
 from .crawler import rate_limiter, requests_with_retry
-from .gameday_dev import get_pbp_data
+from .gameday import get_pbp_data
 
 __all__ = ["rate_limiter", "requests_with_retry", "get_pbp_data"]
